refactor: moved str_starts_ends_contains function
